#22 Post,Comment CRUD with session

+++ front/src/components/PostDetail.vue
@@ -31,16 +31,17 @@
         <div class="bottom">
         <hr>
             <router-link to="/examples/post" tag="button">글목록</router-link>
-            
-            <button v-on:click="updatePost()"> 수정 </button>
-            <button v-on:click="deletePost()"> 삭제 </button>
+            <div v-if="authorized == true">
+                <button v-on:click="updatePost()"> 수정 </button>
+                <button v-on:click="deletePost()"> 삭제 </button>
+            </div>
         </div>
 
         <div class="desc">
         <hr>
         <h5>댓글</h5>
-            <textarea v-model="newComment" placeholder="댓글을 입력하세요." ></textarea>
-            <button v-on:click="createComment()"> 댓글 작성 </button>
+            <textarea v-model="newComment" class="newCommentBox"></textarea>
+            <button v-on:click="createComment()" class="createComment"> 댓글 작성 </button>
             <hr>
             
             <div v-for="comment in comments" class="comments" v-bind:id="comment.id+'test'">
@@ -62,12 +63,12 @@
                     </div>
 
                     <div v-if="updateCommentTF === false">
-                        <button v-on:click="updateTF(comment.id)"> 수정 </button>
-                        <button v-on:click="deleteComment(comment.id)"> 삭제 </button>
+                        <button v-on:click="updateTF()"> 수정 </button>
+                        <button v-on:click="deleteComment(comment)"> 삭제 </button>
                     </div>
                     <div v-else>
-                        <button v-on:click="updateComment(comment.id)"> 수정 완료 </button>
-                        <button v-on:click="updateTF(comment.id)"> 수정 취소 </button>
+                        <button v-on:click="updateComment(comment)"> 수정 완료 </button>
+                        <button v-on:click="updateTF()"> 수정 취소 </button>
                     </div>
                     <hr>
                 </div>
@@ -86,22 +87,23 @@
             this.$http.get(`/api/post/${id}`)
             .then((response) => {
                 this.post = response.data[0];
+
+                this.$http.get(`/api/login/session`)
+                .then((response) => {
+                    this.user = response.data;
+                    
+                    if(this.user.id == this.post.user) {
+                        this.authorized = true;
+                    } else {
+                        this.authorized = false;
+                    }
+                })
             });
 
             this.$http.get(`/api/post/${id}/comment`)
             .then((response) => {
                 this.comments = response.data;
             });
-
-            this.$http.get(`/api/login/session`)
-            .then((response) => {
-                if(!response.data) {
-                    // alert('login필요합니다.');
-                    // this.$router.push({ name: 'login'})
-                } else {
-                    this.user = response.data;
-                }
-            })
         },
         data: function() {
             return {
@@ -110,6 +112,7 @@
                 newComment:'',
                 comments:[],
                 id:0,
+                authorized:false,
                 updateCommentTF:false
             }
         },
@@ -119,23 +122,47 @@
                 this.$router.push({ name: 'postformupdate', params: { id: id }})
             },
             deletePost: function() {
-                var id = this.$route.params.id;
-                this.$http.delete(`/api/post/${id}`)
+                this.$http.get(`/api/login/session`)
                 .then((response) => {
-                    this.$router.push({ name: 'post'})
-                });
+                    this.user = response.data;
+                    
+                    if (this.user) {
+                        if (this.post.user == this.user.id) {
+                            var id = this.$route.params.id;
+                            this.$http.delete(`/api/post/${id}`)
+                            .then((response) => {
+                                this.$router.push({ name: 'post'})
+                            });
+                        } else {
+                            alert('본인이 작성한 글만 삭제 하실수 있습니다.')
+                        }
+                    } else {
+                        alert('사용자 로그인이 필요합니다.');
+                    }
+                    
+                    
+
+                })
             },
             createComment: function() {
-                var pid = this.$route.params.id;
-                var re = 0;
-                var user = this.user;
-                this.$http.post(`/api/post/${pid}/comment`, {root: pid, re: re, user: user, comment:this.newComment})
+                this.$http.get(`/api/login/session`)
                 .then((response) => {
-                    this.$http.get(`/api/post/${pid}/comment`)
-                    .then((response) => {
-                        this.comments = response.data;
-                    });
-                });
+                    this.user = response.data;
+                    if (this.user) {
+                        var pid = this.$route.params.id;
+                        var re = 0;
+                        var user = this.user;
+                        this.$http.post(`/api/post/${pid}/comment`, {root: pid, re: re, user: user, comment:this.newComment})
+                        .then((response) => {
+                            this.$http.get(`/api/post/${pid}/comment`)
+                            .then((response) => {
+                                this.comments = response.data;
+                            });
+                        });
+                    } else {
+                        alert('사용자 로그인이 필요합니다.');
+                    }
+                })
             },
             updateTF: function() {
                 if (this.updateCommentTF === true){
@@ -144,15 +171,25 @@
                     this.updateCommentTF = true;
                 }
             },
-            deleteComment: function(cid) {
+            deleteComment: function(comment) {
+                var cid = comment.id
                 var pid = this.$route.params.id;
-                this.$http.delete(`/api/post/${pid}/comment/${cid}`)
-                .then((response) => {
-                    this.$http.get(`/api/post/${pid}/comment`)
-                    .then((response) => {
-                        this.comments = response.data;
-                    });
-                });
+
+                if (this.user) {
+                    if (comment.user == this.user.id) {
+                        this.$http.delete(`/api/post/${pid}/comment/${cid}`)
+                        .then((response) => {
+                            this.$http.get(`/api/post/${pid}/comment`)
+                            .then((response) => {
+                                this.comments = response.data;
+                            });
+                        });
+                    } else {
+                        alert('본인의 댓글만 삭제 가능합니다.');
+                    }
+                } else {
+                    alert('사용자 로그인이 필요합니다.');
+                }
             },
             updateComment: function() {
                 var pid = this.$route.params.id;
@@ -162,7 +199,6 @@
 
                 this.$http.put(`/api/post/${pid}/comment/${cid}`, {root: id, re: re, user: user, comment:this.newComment})
                 .then((response) => {
-                    // this.$router.push({ name: 'post'})
                 });
             },
             cancleupdate: function() {
@@ -199,7 +235,7 @@
     .desc {
         height:300px;
     }
-    .bottom{
+    .bottom div{
         float:bottom;
     }
     .displayName{
@@ -208,5 +244,30 @@
     }
     .comments div{
         display:inline;
+    }
+    .createComment{
+        width: 80px;
+        height: 70px;
+        background: #868686;
+        color: white;
+        font-size: 15px;
+        font-weight: bold;
+        border-radius: 5px;
+        vertical-align: top;
+    }
+    .newCommentBox{
+        text-rendering: auto;
+        color: initial;
+        width:80%;
+        height: 70px;
+        letter-spacing: normal;
+        word-spacing: normal;
+        text-transform: none;
+        text-indent: 0px;
+        text-shadow: none;
+        display: inline-block;
+        text-align: start;
+        margin: 0em;
+        font: 400 11px system-ui;
     }
 </style>
